+++ steelc/src/input/commands/steelc_command_manager.cpp
@@ -2,8 +2,10 @@
 
 #include <string>
 
+#include "flags/flag_config.h"
 #include "steelc_command.h"
 #include "steelc_commands_impl.h"
+#include "steelc_command_id.h"
 
 const steelc_command* steelc_command_manager::get_command(const std::string& name) const {
 	auto it = commands.find(name);
@@ -15,8 +17,24 @@ const steelc_command* steelc_command_manager::get_command(const std::string& nam
 
 void steelc_command_manager::register_all() {
 	// REGISTER ALL COMMANDS HERE
-	register_command({ "--help", steelc_commands_impl::help_command_handler });
-	register_command({ "--version", steelc_commands_impl::version_command_handler });
-	register_command({ "build", steelc_commands_impl::build_command_handler });
-	register_command({ "project", steelc_commands_impl::project_command_handler });
+	register_command({
+		"--help",
+		steelc_commands_impl::help_command_handler,
+		flag_config::from_command(STEELC_CMD_HELP)
+	});
+	register_command({
+		"--version",
+		steelc_commands_impl::version_command_handler,
+		flag_config::from_command(STEELC_CMD_VERSION)
+	});
+	register_command({
+		"build",
+		steelc_commands_impl::build_command_handler,
+		flag_config::from_command(STEELC_CMD_BUILD)
+	});
+	register_command({
+		"project",
+		steelc_commands_impl::project_command_handler,
+		flag_config::from_command(STEELC_CMD_PROJECT)
+	});
 }