+++ steelc/src/stproj/source_file.cpp
@@ -4,6 +4,7 @@
 #include <stdexcept>
 #include <filesystem>
 #include <string>
+#include <cstdint>
 
 source_file::source_file(const std::string& full_path, const std::string& relative_path) {
 	this->relative_path = relative_path;
@@ -27,4 +28,11 @@ source_file::source_file(const std::string& full_path, const std::string& relati
 
 std::string source_file::name() const {
 	return name_cache;
+}
+
+uint64_t source_file::get_last_modified_time() const {
+	return static_cast<uint64_t>(std::filesystem::last_write_time(full_path).time_since_epoch().count());
+}
+uint64_t source_file::get_size() const {
+	return static_cast<uint64_t>(std::filesystem::file_size(full_path));
 }